+++ editor/imgui/gists/spinner.cpp
@@ -47,7 +47,7 @@ namespace ImGui {
         window->DrawList->AddRectFilled(bb.Min, ImVec2(pos.x + circleStart, bb.Max.y), bg_col);
         window->DrawList->AddRectFilled(bb.Min, ImVec2(pos.x + circleStart*value, bb.Max.y), fg_col);
         
-        const float t = g.Time;
+        const float t = (float)g.Time;
         const float r = size.y / 2;
         const float speed = 1.5f;
         
@@ -87,7 +87,7 @@ namespace ImGui {
         window->DrawList->PathClear();
         
         int num_segments = 30;
-        int start = abs(ImSin(g.Time*1.8f)*(num_segments-5));
+        int start = (int)abs(ImSin((float)g.Time*1.8f)*(num_segments-5));
         
         const float a_min = IM_PI*2.0f * ((float)start) / (float)num_segments;
         const float a_max = IM_PI*2.0f * ((float)num_segments-3) / (float)num_segments;
@@ -96,11 +96,11 @@ namespace ImGui {
         
         for (int i = 0; i < num_segments; i++) {
             const float a = a_min + ((float)i / (float)num_segments) * (a_max - a_min);
-            window->DrawList->PathLineTo(ImVec2(centre.x + ImCos(a+g.Time*8) * radius,
-                                                centre.y + ImSin(a+g.Time*8) * radius));
+            window->DrawList->PathLineTo(ImVec2(centre.x + ImCos(a+(float)g.Time*8) * radius,
+                                                centre.y + ImSin(a+(float)g.Time*8) * radius));
         }
 
-        window->DrawList->PathStroke(color, false, thickness);
+        window->DrawList->PathStroke(color, false, (float)thickness);
         return true;
     }
